Revert "calculate sigmaQOverP in double"
This reverts commit 92527f55caacd93fc1fe5347f64097cbc2b59321.

+++ Core/src/Propagator/detail/PointwiseMaterialInteraction.cpp
@@ -38,11 +38,9 @@ void PointwiseMaterialInteraction::covarianceContributions(
   }
   // TODO just ionisation loss or full energy loss?
   if (energyLoss) {
-    // Squaring float sigmaQoverP was observed to lead to underflow.
-    // Square in double precision and then cast back to float.
-    const double sigmaQoverP =
+    const float sigmaQoverP =
         computeEnergyLossLandauSigmaQOverP(slab, mass, qOverP, absQ);
-    varianceQoverP = static_cast<float>(sigmaQoverP * sigmaQoverP);
+    varianceQoverP = sigmaQoverP * sigmaQoverP;
   }
 }
